only use syscalls on platforms that we're actually testing
