+++ conn_ecn.go
@@ -1,9 +1,11 @@
-// +build !windows
+// +build darwin linux
 
 package quic
 
 import (
 	"errors"
+	"fmt"
+	"net"
 	"syscall"
 	"time"
 
@@ -12,6 +14,27 @@ import (
 )
 
 const ecnMask uint8 = 0x3
+
+func inspectReadBuffer(c net.PacketConn) (int, error) {
+	conn, ok := c.(interface {
+		SyscallConn() (syscall.RawConn, error)
+	})
+	if !ok {
+		return 0, errors.New("doesn't have a SyscallConn")
+	}
+	rawConn, err := conn.SyscallConn()
+	if err != nil {
+		return 0, fmt.Errorf("couldn't get syscall.RawConn: %w", err)
+	}
+	var size int
+	var serr error
+	if err := rawConn.Control(func(fd uintptr) {
+		size, serr = syscall.GetsockoptInt(int(fd), syscall.SOL_SOCKET, syscall.SO_RCVBUF)
+	}); err != nil {
+		return 0, err
+	}
+	return size, serr
+}
 
 type ecnConn struct {
 	ECNCapablePacketConn
